photo_search_app/requirements.txt
numpy~=1.23.5
tensorflow~=2.11.0
tensorflow-hub~=0.13.0
pillow~=9.4.0
Flask~=2.2.2
Flask-Cors~= 3.0.10
scikit-learn~=1.2.2
waitress~=2.1.2
matplotlib~=3.7.1